refactor(tui): review fixes for chat-state extraction
